pyproject.toml
[project]
name = "x-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=2.14.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
